+++ src/App.css
@@ -2,54 +2,3 @@
   width: 100%;
   max-width: 1280px;
 }
-
-.chat-wrapper {
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-  height: calc(100vh - 2rem);
-}
-
-.chat-body {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  flex: 1;
-  padding: 1rem;
-  margin-bottom: 1rem;
-  border-radius: 0.5rem;
-  overflow-y: auto;
-
-  background-color: #f5f5f5;
-}
-
-.message-wrapper {
-  display: flex;
-  flex-direction: column;
-  width: auto;
-  margin-bottom: 0.75rem;
-  align-self: flex-end;
-
-  & .message-body {
-    padding: 0.75rem;
-    margin-bottom: 0.25rem;
-    border-radius: 0.25rem;
-
-    background-color: #c6e2ff;
-  }
-
-  &.ia {
-    align-self: flex-start;
-
-    & .message-body {
-      background-color: #ceffa9;
-    }
-  }
-
-  & .secondary-data {
-    align-self: flex-end;
-    font-size: 70%;
-    color: #666;
-  }
-}
